Port doomgeneric to libpayload
- Replace some library functions not implemented in libpayload with
generally equivalent calls, remove unnecessary ones
- Emulate stdio.h file IO functions with memory buffers, loading from
  CBFS. Only bare minimum to get the game to run, everything else just
returns an error value.
- Currently only works with PS/2 keyboards. Weapon hotkeys not
  currently working.
- Assumes coreboot FB is RGB888, with bits 24:31 as padding

+++ doomgeneric/d_loop.c
@@ -380,7 +380,7 @@ void D_StartNetGame(net_gamesettings_t *settings,
     i = M_CheckParmWithArgs("-extratics", 1);
 
     if (i > 0)
-        settings->extratics = atoi(myargv[i+1]);
+        settings->extratics = atol(myargv[i+1]);
     else
         settings->extratics = 1;
 
@@ -395,7 +395,7 @@ void D_StartNetGame(net_gamesettings_t *settings,
     i = M_CheckParmWithArgs("-dup", 1);
 
     if (i > 0)
-        settings->ticdup = atoi(myargv[i+1]);
+        settings->ticdup = atol(myargv[i+1]);
     else
         settings->ticdup = 1;
